Initial SDL3 Port.

+++ HexaEngine.Core/Windows/Events/DropEventArgs.cs
@@ -5,6 +5,6 @@
         /// <summary>
         /// Gets the timestamp of an event.
         /// </summary>
-        public uint Timestamp { get; set; }
+        public ulong Timestamp { get; set; }
     }
 }
